refactor statements
