+++ src/main/java/com/abc/Customer.java
@@ -9,7 +9,7 @@ import static java.lang.Math.abs;
 
 public class Customer {
     private String name;
-    private List<Account> accounts;
+    private ArrayList<Account> accounts;
 
     public Customer(String name) {
         this.name = name;
@@ -28,36 +28,15 @@ public class Customer {
         return accounts.size();
     }
 
+    public ArrayList<Account> getAccounts() {
+        return accounts;
+    }
+
     public double totalInterestEarned() {
         double total = 0;
         for (Account a : accounts)
             total += a.interestEarned();
         return total;
-    }
-
-    public String getStatement() {
-        String statement = null;
-        statement = "Statement for " + name + "\n";
-        double total = 0.0;
-        for (Account a : accounts) {
-            statement += "\n" + statementForAccount(a) + "\n";
-            total += a.calculateBalance();
-        }
-        statement += "\nTotal In All Accounts " + toDollars(total);
-        return statement;
-    }
-
-    private String statementForAccount(Account a) {//TODO
-        String s = "";
-        s = s + a.getName() + "\n";
-        //Now total up all the transactions
-        double total = 0.0;
-        for (Transaction t : a.getTransactions()) {
-            s += "  " + (t.getAmount() < 0 ? "withdrawal" : "deposit") + " " + toDollars(t.getAmount()) + "\n";
-            total += t.getAmount();
-        }
-        s += "Total " + toDollars(total);
-        return s;
     }
 
     private String toDollars(double d){
